+++ osb-service/src/main/java/de/evoila/cf/broker/custom/mongodb/MongoDBCustomImplementation.java
@@ -10,12 +10,16 @@ import de.evoila.cf.broker.model.Platform;
 import de.evoila.cf.broker.model.ServiceInstance;
 import de.evoila.cf.broker.model.catalog.plan.Plan;
 import de.evoila.cf.broker.model.credential.UsernamePasswordCredential;
+import de.evoila.cf.cpi.bosh.MongoDBBoshPlatformService;
+import de.evoila.cf.cpi.bosh.deployment.manifest.Manifest;
+import de.evoila.cf.security.credentials.CredentialStore;
 import org.bson.Document;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
 import java.util.Map;
 
 /**
@@ -23,13 +27,17 @@ import java.util.Map;
  */
 @Service
 public class MongoDBCustomImplementation {
+    private final CredentialStore credentialStore;
+    private final MongoDBBoshPlatformService boshPlatformService;
 
     private Logger log = LoggerFactory.getLogger(MongoDBCustomImplementation.class);
 
-    private ExistingEndpointBean existingEndpointBean;
+    private final ExistingEndpointBean existingEndpointBean;
 
-    public MongoDBCustomImplementation(ExistingEndpointBean existingEndpointBean) {
+    public MongoDBCustomImplementation(ExistingEndpointBean existingEndpointBean,MongoDBBoshPlatformService boshPlatformService,CredentialStore credentialStore) {
         this.existingEndpointBean = existingEndpointBean;
+        this.boshPlatformService = boshPlatformService;
+        this.credentialStore = credentialStore;
     }
 
     public void createDatabase(MongoDBService connection, String database) throws PlatformException {
@@ -72,12 +80,31 @@ public class MongoDBCustomImplementation {
     public MongoDBService connection(ServiceInstance serviceInstance, Plan plan, UsernamePasswordCredential usernamePasswordCredential) {
         MongoDBService mongoDbService = new MongoDBService();
 
+        Manifest manifest = null;
+        try {
+            manifest = boshPlatformService.getDeployedManifest(serviceInstance);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+        String caCert = null, clientKeyCert = null;
+        try {
+            Object objCa = ((Map<String, Object>) ((Map<String, Object>) manifest.getInstanceGroup("mongodb").get().getProperties().get("mongodb")).get("tls")).get("ca");
+            if(objCa != null)
+                caCert = objCa.toString().trim();
+            if (caCert != null && (caCert.equals("((server_ca))") || caCert.equals("((server_cert.ca))"))) {
+                caCert = credentialStore.getCertificate(serviceInstance.getId(), "server_ca").getCertificate();
+            }
+        }
+        catch (Exception e) {
+            log.warn(e.getMessage());
+        }
+
         if(plan.getPlatform() == Platform.BOSH)
             mongoDbService.createConnection(usernamePasswordCredential.getUsername(), usernamePasswordCredential.getPassword(),
-                    "admin", serviceInstance.getHosts());
+                    "admin", serviceInstance.getHosts(),caCert, null);
         else if (plan.getPlatform() == Platform.EXISTING_SERVICE)
             mongoDbService.createConnection(existingEndpointBean.getUsername(), existingEndpointBean.getPassword(),
-                    existingEndpointBean.getDatabase(), existingEndpointBean.getHosts());
+                    existingEndpointBean.getDatabase(), existingEndpointBean.getHosts(),caCert, null);
 
         return mongoDbService;
     }