+++ osb-service/src/main/java/de/evoila/cf/cpi/CredentialConstants.java
@@ -5,6 +5,8 @@ package de.evoila.cf.cpi;
  */
 public class CredentialConstants {
 
+    public static final String SERVER_CERT = "server_cert";
+    public static final String SERVER_CA = "server_ca";
     public static String ROOT_CREDENTIALS = "root_credentials";
     public static String BACKUP_CREDENTIALS = "backup_credentials";
     public static String BACKUP_AGENT_CREDENTIALS = "backup_agent_credentials";